chore: refactor collector & middleware to retry dir
