+++ momento-sdk/src/intTest/java/momento/sdk/retry/BaseCacheRetryTestClass.java
@@ -5,14 +5,14 @@ import static org.slf4j.LoggerFactory.getLogger;
 import java.time.Duration;
 import java.util.UUID;
 import momento.sdk.CacheClient;
-import momento.sdk.TestRetryMetricsCollector;
-import momento.sdk.TestRetryMetricsMiddleware;
-import momento.sdk.TestRetryMetricsMiddlewareArgs;
 import momento.sdk.auth.CredentialProvider;
 import momento.sdk.config.Configuration;
 import momento.sdk.config.Configurations;
 import momento.sdk.responses.cache.control.CacheCreateResponse;
 import momento.sdk.responses.cache.control.CacheDeleteResponse;
+import momento.sdk.retry.utils.TestRetryMetricsCollector;
+import momento.sdk.retry.utils.TestRetryMetricsMiddleware;
+import momento.sdk.retry.utils.TestRetryMetricsMiddlewareArgs;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.slf4j.Logger;
