+++ momento-sdk/src/intTest/java/momento/sdk/retry/utils/TestRetryMetricsCollector.java
@@ -1,4 +1,4 @@
-package momento.sdk;
+package momento.sdk.retry.utils;
 
 import java.util.ArrayList;
 import java.util.HashMap;
