+++ momento-sdk/src/intTest/java/momento/sdk/retry/utils/TestRetryMetricsMiddleware.java
@@ -1,4 +1,4 @@
-package momento.sdk;
+package momento.sdk.retry.utils;
 
 import momento.sdk.config.middleware.Middleware;
 import momento.sdk.config.middleware.MiddlewareRequestHandler;
