+++ momento-sdk/src/intTest/java/momento/sdk/retry/utils/TestRetryMetricsMiddlewareRequestHandler.java
@@ -1,4 +1,4 @@
-package momento.sdk;
+package momento.sdk.retry.utils;
 
 import io.grpc.Metadata;
 import java.util.concurrent.CompletableFuture;
